Add files via upload
Needed to start with a blank slate to get my head back in the game. I feel like there are going to be looping issues, but i like the way this code is looking. In general it is hard to wrap by head around the code. It would be nice if there was a way to slow down my code line by line and see when each line of code is being executed.