Update createTask_7.js
I did a lot of work in VSC. I realized that a lot of the variable would not be able to be plugged into the functions as they were declared in the getMoveFromUser() function. with the help of chatGPT (:/) I fixed it, and i understand what i did wrong. As well as this issue, I started plugging the js into the HTML so debugging is a lot easier. The updateHTML() is back so that helps a lot. While I did a lot of work today, I still have quit a way to go. Thanks Benji for all your help :D